+++ src/modules/TransientPropagation/TransientPropagationModule.cpp
@@ -72,7 +72,7 @@ TransientPropagationModule::TransientPropagationModule(Configuration& config,
     config_.setDefault<bool>("output_plots_align_pixels", false);
     config_.setDefault<double>("output_plots_theta", 0.0f);
     config_.setDefault<double>("output_plots_phi", 0.0f);
-    config_.setDefault<int>("output_max_gain_histo", 25);
+    config_.setDefault<unsigned int>("output_max_gain_histo", 25);
 
     // Copy some variables from configuration to avoid lookups:
     temperature_ = config_.get<double>("temperature");
@@ -92,7 +92,7 @@ TransientPropagationModule::TransientPropagationModule(Configuration& config,
     output_linegraphs_recombined_ = config_.get<bool>("output_linegraphs_recombined");
     output_linegraphs_trapped_ = config_.get<bool>("output_linegraphs_trapped");
     output_plots_step_ = config_.get<double>("output_plots_step");
-    output_max_gain_histo_ = config.get<int>("output_max_gain_histo");
+    output_max_gain_histo_ = config.get<unsigned int>("output_max_gain_histo");
 
     // Avoids wrong gain histogram inputs
     if(output_max_gain_histo_ <2) {
@@ -329,27 +329,27 @@ void TransientPropagationModule::initialize() {
             gain_primary_histo_ = CreateHistogram<TH1D>(
                 "gain_primary_histo",
                 "Gain per primarily induced charge carrier group after propagation;gain;number of groups transported",
-                output_max_gain_histo_ - 1,
+                static_cast<int>(output_max_gain_histo_ - 1),
                 1,
-                output_max_gain_histo_);
+                static_cast<double>(output_max_gain_histo_));
             gain_all_histo_ =
                 CreateHistogram<TH1D>("gain_all_histo",
                                       "Gain per charge carrier group after propagation;gain;number of groups transported",
-                                      output_max_gain_histo_ - 1,
+                                      static_cast<int>(output_max_gain_histo_ - 1),
                                       1,
-                                      output_max_gain_histo_);
+                                      static_cast<double>(output_max_gain_histo_));
             gain_e_histo_ =
                 CreateHistogram<TH1D>("gain_e_histo",
                                       "Gain per primary electron group after propagation;gain;number of groups transported",
-                                      output_max_gain_histo_ - 1,
+                                      static_cast<int>(output_max_gain_histo_ - 1),
                                       1,
-                                      output_max_gain_histo_);
+                                      static_cast<double>(output_max_gain_histo_));
             gain_h_histo_ =
                 CreateHistogram<TH1D>("gain_h_histo",
                                       "Gain per primary hole group after propagation;gain;number of groups transported",
-                                      output_max_gain_histo_ - 1,
+                                      static_cast<int>(output_max_gain_histo_ - 1),
                                       1,
-                                      output_max_gain_histo_);
+                                      static_cast<double>(output_max_gain_histo_));
             multiplication_level_histo_ = CreateHistogram<TH1D>(
                 "multiplication_level_histo",
                 "Multiplication level of propagated charge carriers;multiplication level;charge carriers",
